backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
Werkzeug==3.0.1
pytest==7.4.0
pymysql==1.1.0
SQLAlchemy==2.0.23
cryptography==41.0.4
flask-cors==4.0.0


